fix(mail-with-thread): quick changes
