+++ server/api/chat/utils.ts
@@ -1,4 +1,4 @@
-import { splitGroupedCitationsWithSpaces } from "@/utils"
+import { splitGroupedCitationsWithSpaces, getErrorMessage } from "@/utils"
 import {
   Apps,
   CalendarEntity,
@@ -41,7 +41,173 @@ import {
   type AgentReasoningStep,
 } from "@/shared/types"
 import type { Citation } from "@/api/chat/types"
+import { SearchEmailThreads } from "@/search/vespa"
+import { getLoggerWithChild } from "@/logger"
+import type { Span } from "@/tracer"
+import { Subsystem } from "@/types"
 const { maxValidLinks } = config
+
+// Interface for email search result fields
+export interface EmailSearchResultFields {
+  app: Apps
+  threadId?: string
+  docId: string
+  [key: string]: any // Allow other fields
+}
+
+// Type for VespaSearchResult with email fields
+export type EmailSearchResult = VespaSearchResult & {
+  fields: EmailSearchResultFields
+}
+
+export async function expandEmailThreadsInResults(
+  results: VespaSearchResult[],
+  email: string,
+  span?: Span
+): Promise<VespaSearchResult[]> {
+  // Extract unique thread IDs from email results
+  const threadIds = extractThreadIdsFromResults(results)
+  if(threadIds.length === 0){
+    return results;
+  }
+  const { mergedResults, addedCount } = await mergeThreadResults(results, threadIds, email, span)
+  
+  if (addedCount > 0) {
+    getLoggerWithChild(Subsystem.Chat)({ email }).info(
+      `Added ${addedCount} additional emails from ${threadIds.length} threads to search results`
+    )
+  }
+  
+  return mergedResults
+}
+
+// Helper function to process thread results and merge them with existing results
+export function processThreadResults(
+  threadResults: VespaSearchResult[],
+  existingDocIds: Set<string>,
+  mergedResults: VespaSearchResult[]
+): { addedCount: number; threadInfo: Record<string, number> } {
+  let addedCount = 0
+  let threadInfo: Record<string, number> = {}
+  
+  for (const child of threadResults) {
+    const emailChild = child as EmailSearchResult
+    const docId = emailChild.fields.docId
+    const threadId = emailChild.fields.threadId
+    
+    // Skip if already in results
+    if (!existingDocIds.has(docId)) {
+      mergedResults.push(child)
+      existingDocIds.add(docId)
+      addedCount++
+      
+      // Track count per thread for logging
+      if (threadId) {
+        threadInfo[threadId] = (threadInfo[threadId] || 0) + 1
+      }
+    }
+  }
+  
+  return { addedCount, threadInfo }
+}
+
+// Function to extract thread IDs from search results
+export function extractThreadIdsFromResults(results: VespaSearchResult[]): string[] {
+  const seenThreadIds = new Set<string>()
+  
+  return results.reduce<string[]>((threadIds, result) => {
+    const fields = result.fields as EmailSearchResultFields
+    // Check if it's an email result
+    if (fields.app === Apps.Gmail && fields.threadId) {
+      if (!seenThreadIds.has(fields.threadId)) {
+        threadIds.push(fields.threadId)
+        seenThreadIds.add(fields.threadId)
+      }
+    }
+    return threadIds
+  }, [])
+}
+
+// Helper function to merge thread results into existing results
+export async function mergeThreadResults(
+  existingResults: VespaSearchResult[],
+  threadIds: string[],
+  email: string,
+  span?: Span
+): Promise<{ mergedResults: VespaSearchResult[], addedCount: number, threadInfo: Record<string, number> }> {
+  if (threadIds.length === 0) {
+    return { mergedResults: existingResults, addedCount: 0, threadInfo: {} }
+  }
+
+  const threadSpan = span?.startSpan("fetch_email_threads")
+  threadSpan?.setAttribute("threadIds", JSON.stringify(threadIds))
+  
+  try {
+    const threadResults = await SearchEmailThreads(threadIds, email)
+    
+    if (!threadResults.root.children || threadResults.root.children.length === 0) {
+      threadSpan?.setAttribute("no_thread_results", true)
+      threadSpan?.end()
+      return { mergedResults: existingResults, addedCount: 0, threadInfo: {} }
+    }
+    
+    // Create a set of existing docIds to avoid duplicates
+    const existingDocIds = new Set(
+      existingResults.map((child: any) => child.fields.docId)
+    )
+    
+    // Merge thread results
+    const mergedResults = [...existingResults]
+    
+    const { addedCount, threadInfo } = processThreadResults(
+      threadResults.root.children,
+      existingDocIds,
+      mergedResults
+    )
+    
+    threadSpan?.setAttribute("added_email_count", addedCount)
+    threadSpan?.setAttribute("total_thread_emails_found", threadResults.root.children.length)
+    threadSpan?.setAttribute("thread_info", JSON.stringify(threadInfo))
+    threadSpan?.end()
+    
+    return { mergedResults, addedCount, threadInfo }
+  } catch (error) {
+    getLoggerWithChild(Subsystem.Chat)({ email }).error(
+      error,
+      `Error fetching email threads: ${getErrorMessage(error)}`
+    )
+    threadSpan?.setAttribute("error", getErrorMessage(error))
+    threadSpan?.end()
+    return { mergedResults: existingResults, addedCount: 0, threadInfo: {} }
+  }
+}
+
+export const extractImageFileNames = (
+  context: string,
+): { imageFileNames: string[] } => {
+  // This matches "Image File Names:" followed by content until the next field (starting with a capital letter and colon) or "vespa relevance score"
+  const imageContentRegex =
+    /Image File Names:\s*([\s\S]*?)(?=\n[A-Z][a-zA-Z ]*:|vespa relevance score|$)/g
+  const matches = [...context.matchAll(imageContentRegex)]
+
+  let imageFileNames: string[] = []
+
+  for (const match of matches) {
+    const imageContent = match[1].trim()
+    if (imageContent) {
+      // Split by newlines and filter out empty strings
+      const fileNames = imageContent
+        .split("\n")
+        .map((name) => name.trim())
+        .filter((name) => name.length > 0)
+        // Additional safety: split by spaces and filter out empty strings
+        // in case multiple filenames are on the same line
+        .flatMap((name) => name.split(/\s+/).filter((part) => part.length > 0))
+      imageFileNames.push(...fileNames)
+    }
+  }
+  return { imageFileNames }
+}
 
 export const searchToCitation = (result: VespaSearchResults): Citation => {
   const fields = result.fields
